modified sim.xml and macros.xml to allow for system thermocoupling and option to start velocities at NVT (#138)
* modified sim.xml and macros.xml to allow for system thermocoupling and option to start velocities at NVT

* Update tools/gromacs/sim.xml

Co-authored-by: Simon Bray <32272674+simonbray@users.noreply.github.com>

* Update macros.xml

removed the max # of md_steps since currently the 1 million step max only allows for 2ns of simulation time at 2 femtosecond time steps. Also changed the 'help' statement to MD steps, since this is used in both NVT and NPT.

* Update macros.xml

* Update macros.xml

* Update sim.xml

added a conditional for the ensemble selection, which includes whether or not to add velocities only when NVT is selected.

* Update sim.xml

changed "tcouple" to "tcoupl". This was causing the tool to fail lint and testing.

* Update sim.xml

changed "#if $sets.ensemble == "nvt":" to "#if $sets.ensembleselect.ensemble == "nvt":" in the pressure coupling section of the mdp settings, so it's able to find the parameter now. this was causing a planemo test fail. also changed a couple minor spacing issues.

* Update macros.xml

removed </yield>

* Update macros.xml

fixed indentation

Co-authored-by: Simon Bray <32272674+simonbray@users.noreply.github.com>